The 100 commit 1.13 update

+++ plugin/src/main/java/net/aufdemrand/denizen/tags/core/ServerTags.java
@@ -16,6 +16,7 @@ import net.aufdemrand.denizen.scripts.commands.server.BossBarCommand;
 import net.aufdemrand.denizen.scripts.containers.core.AssignmentScriptContainer;
 import net.aufdemrand.denizen.tags.BukkitTagContext;
 import net.aufdemrand.denizen.utilities.DenizenAPI;
+import net.aufdemrand.denizen.utilities.Utilities;
 import net.aufdemrand.denizen.utilities.debugging.dB;
 import net.aufdemrand.denizen.utilities.depends.Depends;
 import net.aufdemrand.denizencore.DenizenCore;
@@ -36,7 +37,6 @@ import net.citizensnpcs.api.npc.NPC;
 import org.bukkit.*;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
-import org.bukkit.event.Listener;
 import org.bukkit.plugin.Plugin;
 
 import java.io.File;
@@ -478,8 +478,7 @@ public class ServerTags {
         if (attribute.startsWith("has_file") && attribute.hasContext(1)) {
             File f = new File(DenizenAPI.getCurrentInstance().getDataFolder(), attribute.getContext(1));
             try {
-                if (!Settings.allowStrangeYAMLSaves() &&
-                        !f.getCanonicalPath().startsWith(DenizenAPI.getCurrentInstance().getDataFolder().getCanonicalPath())) {
+                if (!Utilities.canReadFile(f)) {
                     dB.echoError("Invalid path specified. Invalid paths have been denied by the server administrator.");
                     return;
                 }
@@ -500,13 +499,13 @@ public class ServerTags {
         // -->
         if (attribute.startsWith("list_files") && attribute.hasContext(1)) {
             File folder = new File(DenizenAPI.getCurrentInstance().getDataFolder(), attribute.getContext(1));
-            if (!folder.exists() || !folder.isDirectory()) {
-                return;
-            }
             try {
-                if (!Settings.allowStrangeYAMLSaves() &&
-                        !folder.getCanonicalPath().startsWith(DenizenAPI.getCurrentInstance().getDataFolder().getCanonicalPath())) {
+                if (!Utilities.canReadFile(folder)) {
                     dB.echoError("Invalid path specified. Invalid paths have been denied by the server administrator.");
+                    return;
+                }
+                if (!folder.exists() || !folder.isDirectory()) {
+                    dB.echoError("Invalid path specified. No directory exists at that path.");
                     return;
                 }
             }
@@ -1098,7 +1097,7 @@ public class ServerTags {
         // -->
         else if (attribute.startsWith("entity_is_spawned")
                 && attribute.hasContext(1)) {
-            dEntity ent = dEntity.valueOf(attribute.getContext(1));
+            dEntity ent = dEntity.valueOf(attribute.getContext(1), new BukkitTagContext(null, null, false, null, false, null));
             event.setReplaced(new Element((ent != null && ent.isUnique() && ent.isSpawned()) ? "true" : "false")
                     .getAttribute(attribute.fulfill(1)));
         }
@@ -1123,7 +1122,7 @@ public class ServerTags {
         // -->
         else if (attribute.startsWith("npc_is_valid")
                 && attribute.hasContext(1)) {
-            dNPC npc = dNPC.valueOf(attribute.getContext(1));
+            dNPC npc = dNPC.valueOf(attribute.getContext(1), new BukkitTagContext(null, null, false, null, false, null));
             event.setReplaced(new Element((npc != null && npc.isValid()))
                     .getAttribute(attribute.fulfill(1)));
         }
@@ -1189,6 +1188,10 @@ public class ServerTags {
                 return;
             }
             File file = new File(DenizenAPI.getCurrentInstance().getDataFolder(), value.asString());
+            if (!Utilities.isSafeFile(file)) {
+                dB.echoError("Cannot delete that file (unsafe path).");
+                return;
+            }
             try {
                 if (!file.delete()) {
                     dB.echoError("Failed to delete file: returned false");
